Chage method float to double

+++ src/Notas.java
@@ -1,13 +1,13 @@
 public class Notas {
     //Atributos de cada materia
-    public float notaMateria1;
-    public float notaMAteria2;
-    public float notaMAteria3;
-    public float notaMAteria4;
-    public float notaMAteria5;
+    public double notaMateria1;
+    public double notaMAteria2;
+    public double notaMAteria3;
+    public double notaMAteria4;
+    public double notaMAteria5;
     //Costructor --> Inicializa la clase
 
-    public Notas(float notaMateria1, float notaMAteria2, float notaMAteria3, float notaMAteria4, float notaMAteria5) {
+    public Notas(double notaMateria1, double notaMAteria2, double notaMAteria3, double notaMAteria4, double notaMAteria5) {
         this.notaMateria1 = notaMateria1;
         this.notaMAteria2 = notaMAteria2;
         this.notaMAteria3 = notaMAteria3;
@@ -16,44 +16,44 @@ public class Notas {
     }
     //Getters
 
-    public float getNotaMateria1() {
+    public double getNotaMateria1() {
         return notaMateria1;
     }
 
-    public float getNotaMAteria2() {
+    public double getNotaMAteria2() {
         return notaMAteria2;
     }
 
-    public float getNotaMAteria3() {
+    public double getNotaMAteria3() {
         return notaMAteria3;
     }
 
-    public float getNotaMAteria4() {
+    public double getNotaMAteria4() {
         return notaMAteria4;
     }
 
-    public float getNotaMAteria5() {
+    public double getNotaMAteria5() {
         return notaMAteria5;
     }
     //Setteres
 
-    public void setNotaMateria1(float notaMateria1) {
+    public void setNotaMateria1(double notaMateria1) {
         this.notaMateria1 = notaMateria1;
     }
 
-    public void setNotaMAteria2(float notaMAteria2) {
+    public void setNotaMAteria2(double notaMAteria2) {
         this.notaMAteria2 = notaMAteria2;
     }
 
-    public void setNotaMAteria3(float notaMAteria3) {
+    public void setNotaMAteria3(double notaMAteria3) {
         this.notaMAteria3 = notaMAteria3;
     }
 
-    public void setNotaMAteria4(float notaMAteria4) {
+    public void setNotaMAteria4(double notaMAteria4) {
         this.notaMAteria4 = notaMAteria4;
     }
 
-    public void setNotaMAteria5(float notaMAteria5) {
+    public void setNotaMAteria5(double notaMAteria5) {
         this.notaMAteria5 = notaMAteria5;
     }
 }
